BACKENED/requirements.txt
# ==============================
# 🚀 FASTAPI CORE
# ==============================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ==============================
# 🗄 DATABASE
# ==============================
pymongo==4.6.3
dnspython==2.6.1
bson==0.5.10

# ==============================
# 🔐 AUTH & SECURITY
# ==============================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.1

# ==============================
# ⚙️ CONFIG MANAGEMENT
# ==============================
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ==============================
# 🤖 AI / LLM (Optional but Recommended)
# ==============================
openai==1.23.2
tiktoken==0.6.0

# ==============================
# 📊 UTILITIES
# ==============================
python-multipart==0.0.9
requests==2.31.0

# ==============================
# 🧪 DEVELOPMENT (Optional)
# ==============================
pytest==8.1.1
httpx==0.27.0